refactor: 게임 로직 및 이미지 익스포터 모듈화
프로젝트의 모듈성을 개선하기 위한 리팩토링을 계속 진행했습니다. 이번에는 게임 모드 로직과 이미지 저장 기능의 구조를 개선했습니다.

- gameMode.js 리팩토링: 중복되던 카드 수집 로직을 getAllCardsInRoster 헬퍼 함수로 추출하여 코드의 가독성과 효율성을 높였습니다.

- imageExporter.js 분해: 하나의 파일이 너무 많은 역할을 하던 imageExporter.js를 기능에 따라 분리했습니다.

- 확인 모달창(UI)은 confirmationModal.js로, HTML 생성 로직은 rosterHtmlGenerator.js로 분리하여 각 모듈이 하나의 책임만 갖도록 했습니다.
