+++ modules/gameMode.js
@@ -17,6 +17,28 @@ const cardStatusTransitions = {
     }
 };
 
+const getAllCardsInRoster = (roster) => {
+    const allCards = [];
+    if (!roster) return allCards;
+
+    if (roster.units) {
+        Object.values(roster.units).forEach(unit => allCards.push(...Object.values(unit)));
+    }
+    if (roster.drones) {
+        allCards.push(...roster.drones);
+        roster.drones.forEach(drone => {
+            if (drone && drone.backCard) {
+                allCards.push(drone.backCard);
+            }
+        });
+    }
+    if (roster.tacticalCards) {
+        allCards.push(...roster.tacticalCards);
+    }
+    
+    return allCards.filter(Boolean); // Filter out any null/undefined cards
+};
+
 export function advanceCardStatus(card, unit) {
     if (!card) return;
 
@@ -72,12 +94,10 @@ export function performActionAndPreserveScroll(action, eventTarget) {
 }
 
 const initializeSubDrones = (gameRoster) => {
-    const allUnitAndDroneCards = [];
-    Object.values(gameRoster.units).forEach(unit => allUnitAndDroneCards.push(...Object.values(unit)));
-    allUnitAndDroneCards.push(...gameRoster.drones);
-
+    const allCards = getAllCardsInRoster(gameRoster);
     const processedSubDrones = new Set(gameRoster.drones.map(d => d.fileName));
-    allUnitAndDroneCards.forEach(card => {
+
+    allCards.forEach(card => {
         if (card && card.subCards) {
             card.subCards.forEach(subCardFileName => {
                 const subCardData = state.allCards.byFileName.get(subCardFileName);
@@ -93,17 +113,7 @@ const initializeSubDrones = (gameRoster) => {
 };
 
 const initializeCardStates = (gameRoster) => {
-    const allCardsInGame = [];
-    Object.values(gameRoster.units).forEach(unit => allCardsInGame.push(...Object.values(unit)));
-    allCardsInGame.push(...gameRoster.drones);
-    if (gameRoster.tacticalCards) {
-        allCardsInGame.push(...gameRoster.tacticalCards);
-    }
-    gameRoster.drones.forEach(drone => {
-        if (drone.backCard) {
-            allCardsInGame.push(drone.backCard);
-        }
-    });
+    const allCardsInGame = getAllCardsInRoster(gameRoster);
 
     allCardsInGame.forEach(card => {
         if (!card) return;
